requirements.txt
djangorestframework
Django~=5.2.7
